+++ e2e/cases/inspect-config/debug.test.ts
@@ -1,6 +1,7 @@
 import path from 'node:path';
 import { build, dev, gotoPage } from '@e2e/helper';
 import { expect, test } from '@playwright/test';
+import { logger } from '@rsbuild/core';
 import { fse } from '@rsbuild/shared';
 
 const getRsbuildConfig = (dist: string) =>
@@ -12,7 +13,7 @@ const getBundlerConfig = (dist: string) =>
   );
 
 test('should generate config files when build (with DEBUG)', async () => {
-  process.env.DEBUG = 'rsbuild';
+  logger.level = 'verbose';
   const distRoot = 'dist-1';
 
   await build({
@@ -30,11 +31,11 @@ test('should generate config files when build (with DEBUG)', async () => {
   expect(fse.existsSync(getRsbuildConfig(distRoot))).toBeTruthy();
   expect(fse.existsSync(getBundlerConfig(distRoot))).toBeTruthy();
 
-  delete process.env.DEBUG;
+  logger.level = 'log';
 });
 
 test('should generate config files when dev (with DEBUG)', async ({ page }) => {
-  process.env.DEBUG = 'rsbuild';
+  logger.level = 'verbose';
   const distRoot = 'dist-2';
 
   const rsbuild = await dev({
@@ -56,7 +57,7 @@ test('should generate config files when dev (with DEBUG)', async ({ page }) => {
   expect(fse.existsSync(getRsbuildConfig(distRoot))).toBeTruthy();
   expect(fse.existsSync(getBundlerConfig(distRoot))).toBeTruthy();
 
-  delete process.env.DEBUG;
+  logger.level = 'log';
 
   await rsbuild.close();
 });
