+++ packages/compat/webpack/src/module.d.ts
@@ -1,0 +1,1 @@
+declare module 'webpack/lib/MultiStats.js';
