+++ packages/core/src/internal.ts
@@ -12,10 +12,11 @@ export { initRsbuildConfig } from './provider/initConfigs';
 export { stringifyConfig, outputInspectConfigFiles } from './config';
 export type { InternalContext } from './types';
 export { setHTMLPlugin, getHTMLPlugin } from './pluginHelper';
-export { formatStats, getStatsOptions } from './helpers';
+export { formatStats, getStatsOptions, onCompileDone } from './helpers';
 export { getChainUtils } from './provider/rspackConfig';
 export { chainToConfig, modifyBundlerChain } from './configChain';
 export { applySwcDecoratorConfig } from './plugins/swc';
 export { getSwcMinimizerOptions } from './plugins/minimize';
 export { getDevMiddleware } from './server/devMiddleware';
 export { createDevServer } from './server/devServer';
+export { TARGET_ID_MAP } from './constants';
