+++ packages/core/src/pluginManager.ts
@@ -1,4 +1,4 @@
-import { color, debug, isFunction, logger } from '@rsbuild/shared';
+import { color, isFunction, logger } from '@rsbuild/shared';
 import type {
   BundlerPluginInstance,
   Falsy,
@@ -173,7 +173,7 @@ export async function initPlugins({
   pluginAPI?: RsbuildPluginAPI;
   pluginManager: PluginManager;
 }) {
-  debug('init plugins');
+  logger.debug('init plugins');
 
   const plugins = pluginDagSort(pluginManager.getPlugins());
 
@@ -191,5 +191,5 @@ export async function initPlugins({
     await plugin.setup(pluginAPI!);
   }
 
-  debug('init plugins done');
+  logger.debug('init plugins done');
 }
