+++ packages/core/src/plugins/entry.ts
@@ -3,10 +3,10 @@ import {
   type RsbuildTarget,
   castArray,
   color,
-  createVirtualModule,
   reduceConfigsMergeContext,
 } from '@rsbuild/shared';
 import type { EntryDescription } from '@rspack/core';
+import { createVirtualModule } from '../helpers';
 import type { NormalizedConfig, RsbuildConfig, RsbuildPlugin } from '../types';
 
 export function getEntryObject(
