+++ packages/core/src/rspack/HtmlBasicPlugin.ts
@@ -5,12 +5,11 @@ import {
   type HtmlTagUtils,
   type ModifyHTMLTagsFn,
   isFunction,
-  partition,
 } from '@rsbuild/shared';
 import type { Compilation, Compiler } from '@rspack/core';
 import type HtmlWebpackPlugin from 'html-webpack-plugin';
 import type { HtmlTagObject } from 'html-webpack-plugin';
-import { ensureAssetPrefix } from '../helpers';
+import { ensureAssetPrefix, partition } from '../helpers';
 import { getHTMLPlugin } from '../pluginHelper';
 
 export type TagConfig = {
