+++ packages/core/src/server/middlewares.ts
@@ -5,7 +5,6 @@ import {
   type RequestHandler as Middleware,
   type Rspack,
   color,
-  debug,
   isDebug,
   logger,
 } from '@rsbuild/shared';
@@ -57,7 +56,7 @@ export const getRequestLoggerMiddleware: () => Promise<Connect.NextHandleFunctio
           (endAt[0] - _startAt[0]) * 1e3 + (endAt[1] - _startAt[1]) * 1e-6;
 
         // :status :method :url :total-time ms
-        debug(
+        logger.debug(
           `${statusColor(status)} ${method} ${color.gray(url)} ${color.gray(
             `${totalTime.toFixed(3)} ms`,
           )}`,
@@ -127,7 +126,7 @@ export const getHtmlFallbackMiddleware: (params: {
 
     const rewrite = (newUrl: string, isFallback = false) => {
       if (isFallback && isDebug()) {
-        debug(
+        logger.debug(
           `${req.method} ${color.gray(
             `${req.url} ${color.yellow('fallback')} to ${newUrl}`,
           )}`,
