+++ packages/core/tests/helpers.test.ts
@@ -1,5 +1,15 @@
-import { ensureAssetPrefix } from '../src/helpers';
+import { ensureAssetPrefix, pick } from '../src/helpers';
 import { normalizeUrl } from '../src/server/helper';
+
+describe('pick', () => {
+  it('should pick from object correctly', () => {
+    expect(pick({ foo: 1, bar: 2 }, ['foo'])).toEqual({ foo: 1 });
+  });
+
+  it('should not pick undefined properties', () => {
+    expect(pick({ foo: undefined, bar: undefined }, ['foo'])).toEqual({});
+  });
+});
 
 it('normalizeUrl', () => {
   expect(normalizeUrl('https://www.example.com/static//a')).toBe(
