+++ packages/shared/src/constants.ts
@@ -1,5 +1,3 @@
-import type { RsbuildTarget } from './types';
-
 const DEFAULT_WEB_BROWSERSLIST = [
   'chrome >= 87',
   'edge >= 88',
@@ -22,10 +20,3 @@ export const TS_REGEX = /\.(?:ts|mts|cts|tsx)$/;
 export const SCRIPT_REGEX = /\.(?:js|jsx|mjs|cjs|ts|tsx|mts|cts)$/;
 export const TS_AND_JSX_REGEX = /\.(?:ts|tsx|jsx|mts|cts)$/;
 export const NODE_MODULES_REGEX = /[\\/]node_modules[\\/]/;
-
-export const TARGET_ID_MAP: Record<RsbuildTarget, string> = {
-  web: 'Client',
-  node: 'Server',
-  'web-worker': 'Web Worker',
-  'service-worker': 'Service Worker',
-};
