+++ packages/shared/src/index.ts
@@ -1,6 +1,9 @@
+import fse from '../compiled/fs-extra/index.js';
+
+export { fse };
+
 export * from './constants';
 export * from './utils';
-export * from './fs';
 export * from './getBrowserslist';
 export * from './logger';
 export * from './reduceConfigs';
