+++ packages/shared/tests/utils.test.ts
@@ -1,26 +1,4 @@
-import { camelCase, createDependenciesRegExp } from '../src';
-
-describe('camelCase', () => {
-  test('should convert snake_case to camelCase', () => {
-    expect(camelCase('hello_world')).toBe('helloWorld');
-  });
-
-  test('should convert kebab-case to camelCase', () => {
-    expect(camelCase('foo-bar-baz')).toBe('fooBarBaz');
-  });
-
-  test('should handle already camelCase input', () => {
-    expect(camelCase('alreadyCamelCase')).toBe('alreadyCamelCase');
-  });
-
-  test('should handle empty input', () => {
-    expect(camelCase('')).toBe('');
-  });
-
-  test('should handle single-word input', () => {
-    expect(camelCase('single')).toBe('single');
-  });
-});
+import { createDependenciesRegExp } from '../src';
 
 test('createDependenciesRegExp', () => {
   const cases = {
